LayoutTests/editing/undo-manager/undo-manager-keeps-wrapper-alive.html

<!DOCTYPE html> <!-- webkit-test-runner [ enableUndoManagerAPI=true ] -->
<html>
    <head>
        <script src="../../resources/js-test.js"></script>
    </head>
    <script>
        function runTest()
        {
            description("Verifies that the wrapper for Document's UndoManager remains consistent across GC passes. This test must be run using WebKitTestRunner.");
            document.undoManager.foo = "bar";
            if (window.GCController)
                GCController.collect();
            shouldBeEqualToString("document.undoManager.foo", "bar");
        }
    </script>
    <body onload=runTest()></body>
</html>
