LayoutTests/editing/undo-manager/undo-manager-keeps-wrapper-alive-expected.txt

Verifies that the wrapper for Document's UndoManager remains consistent across GC passes. This test must be run using WebKitTestRunner.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.undoManager.foo is "bar"
PASS successfullyParsed is true

TEST COMPLETE
